Dockerfiles: support multi-arch + improvements
Allow building multi-arch images. The user can now use the `--platform`
command line parameter, to build different platform.

The builder image runs on the local machine platform, to acheive better
performance, But the final image is build according to the user request.

Bump go to 1.23

The final image does not need go. replace it's base image to plain
alpine, instead of golang alpine.

re-order the Dockerfile commands to get better caching, and fewer new
layers when storing the image in the registry.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>